+++ src/Query.java
@@ -1,19 +1,19 @@
 public class Query {
 
-    int[][] matrix;
+    float[][] matrix;
 
 
-    Query(int[][] matrix) {
+    Query(float[][] matrix) {
         this.matrix = matrix;
     }
 
 
-    int getWeight(int a, int b) {
+    float getWeight(int a, int b) {
         return matrix[a][b];
     }
 
-    int getObjectiveValue(int path[]) {
-        int cost = 0;
+    float getObjectiveValue(int path[]) {
+        float cost = 0;
 
         for (int i = 0; i < path.length - 1; i++) {
             cost += matrix[path[i]][path[i + 1]];
